requirements.txt
streamlit==1.32.0
google-generativeai==0.4.1
firebase-admin==6.5.0
pillow==10.2.0
